refactor(parser-jvm-plugin-backbone): use default ObjectMapper
