+++ packages/java/parser-jvm-plugin-backbone/src/test/java/dev/hilla/parser/plugins/backbone/generics/GenericsBareRefEntity.java
@@ -2,4 +2,8 @@ package dev.hilla.parser.plugins.backbone.generics;
 
 public class GenericsBareRefEntity<T> {
     private T bareGenericTypeProperty;
+
+    public T getBareGenericTypeProperty() {
+        return bareGenericTypeProperty;
+    }
 }
