+++ packages/java/parser-jvm-plugin-backbone/src/test/java/dev/hilla/parser/plugins/backbone/generics/GenericsExtendedRefEntity.java
@@ -2,4 +2,8 @@ package dev.hilla.parser.plugins.backbone.generics;
 
 public class GenericsExtendedRefEntity<T extends GenericsBareRefEntity<String>> {
     private T extendedGenericTypeReference;
+
+    public T getExtendedGenericTypeReference() {
+        return extendedGenericTypeReference;
+    }
 }
