+++ packages/java/parser-jvm-plugin-backbone/src/test/java/dev/hilla/parser/plugins/backbone/multiendpoints/MultiEndpointsSharedModel.java
@@ -2,4 +2,8 @@ package dev.hilla.parser.plugins.backbone.multiendpoints;
 
 public class MultiEndpointsSharedModel {
     private String id;
+
+    public String getId() {
+        return id;
+    }
 }
