+++ packages/java/parser-jvm-plugin-backbone/src/test/java/dev/hilla/parser/plugins/backbone/shadowedname/subpackage/Collection.java
@@ -5,4 +5,12 @@ import java.util.List;
 public class Collection<T> {
     private List<T> items;
     private String name;
+
+    public List<T> getItems() {
+        return items;
+    }
+
+    public String getName() {
+        return name;
+    }
 }
